# Switch from mocha to ava to make tests faster