--- a/.gitignore
+++ b/.gitignore
@@ -10,3 +10,4 @@ tsconfig.schemastore-schema.json
 .idea/
 /.vscode/
 /website/static/api
+/tsconfig.tsbuildinfo