--- a/.mocharc.js
+++ b/.mocharc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  spec: ['dist/**/*.spec.js']
-};