--- a/package.json
+++ b/package.json
@@ -51,8 +51,8 @@
     "build-tsc": "tsc",
     "build-configSchema": "typescript-json-schema --topRef --refs --validationKeywords allOf --out tsconfig.schema.json tsconfig.json TsConfigSchema && node --require ./register ./scripts/create-merged-schema",
     "build-pack": "node ./scripts/build-pack.js",
-    "test-spec": "mocha dist/**/*.spec.js -R spec --bail",
-    "test-cov": "nyc mocha -- \"dist/**/*.spec.js\" -R spec --bail",
+    "test-spec": "ava",
+    "test-cov": "nyc ava",
     "test": "npm run build && npm run lint && npm run test-cov --",
     "coverage-report": "nyc report --reporter=lcov",
     "prepare": "npm run build-nopack"
@@ -89,22 +89,28 @@
     "url": "https://github.com/TypeStrong/ts-node/issues"
   },
   "homepage": "https://github.com/TypeStrong/ts-node",
+  "ava": {
+    "files": [
+      "dist/*.spec.js"
+    ],
+    "failWithoutAssertions": false,
+    "timeout": "100s"
+  },
   "devDependencies": {
     "@types/chai": "^4.0.4",
     "@types/diff": "^4.0.2",
     "@types/lodash": "^4.14.151",
-    "@types/mocha": "^5.2.7",
     "@types/node": "13.13.5",
     "@types/proxyquire": "^1.3.28",
     "@types/react": "^16.0.2",
     "@types/rimraf": "^3.0.0",
     "@types/semver": "^7.1.0",
     "@types/source-map-support": "^0.5.0",
+    "ava": "^3.15.0",
     "axios": "^0.21.1",
     "chai": "^4.0.1",
     "get-stream": "^6.0.0",
     "lodash": "^4.17.15",
-    "mocha": "^6.2.3",
     "ntypescript": "^1.201507091536.1",
     "nyc": "^15.0.1",
     "proxyquire": "^2.0.0",