--- a/package.json
+++ b/package.json
@@ -117,6 +117,7 @@
     "react": "^16.14.0",
     "rimraf": "^3.0.0",
     "semver": "^7.1.3",
+    "throat": "^6.0.1",
     "tslint": "^6.1.0",
     "tslint-config-standard": "^9.0.0",
     "typedoc": "^0.20.20",