--- a/package.json
+++ b/package.json
@@ -57,7 +57,7 @@
     "test": "npm run build && npm run lint && npm run test-cov --",
     "test-local": "npm run lint && npm run build-tsc && npm run build-pack && npm run test-spec --",
     "coverage-report": "nyc report --reporter=lcov",
-    "prepare": "npm run build-nopack"
+    "prepare": "npm run clean && npm run build-nopack"
   },
   "engines": {
     "node": ">=10.0.0"