--- a/src/index.ts
+++ b/src/index.ts
@@ -1,3 +1,4 @@
+import { test } from './testlib'
 import { relative, basename, extname, resolve, dirname, join } from 'path'
 import sourceMapSupport = require('source-map-support')
 import * as ynModule from 'yn'