--- a/tests/env.ts
+++ b/tests/env.ts
@@ -1,1 +1,1 @@
-console.log(typeof process[Symbol.for('ts-node.register.instance')])
+console.log(typeof (process as any)[Symbol.for('ts-node.register.instance')])