--- a/tests/tsconfig.json
+++ b/tests/tsconfig.json
@@ -3,6 +3,7 @@
     "target": "es2015",
     "jsx": "react",
     "noEmit": true,
+    "strict": true,
     // Global type definitions.
     "typeRoots": [
       "./typings",