--- a/tsconfig.build-schema.json
+++ b/tsconfig.build-schema.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "incremental": false
+  }
+}