--- a/tsconfig.json
+++ b/tsconfig.json
@@ -11,7 +11,7 @@
     "declaration": true,
     "sourceMap": true,
     "inlineSources": true,
-    "types": ["node", "mocha"],
+    "types": ["node"],
     "stripInternal": true
   },
   "include": [