`npm test` with ava: 1m 50s
`npm test` with mocha: 2m 47s

`npm run test-spec` with ava: 55s
`npm run test-spec` with mocha: 1m 56s

Ignoring how slow the build is -- I should try to fix that in another PR -- ava chops a minute off of test execution time.

## Other things snuck in

* turn on tsc "incremental" and "skipLibCheck" to make builds faster
* add `npm run test-local` which is faster way to run tests in local development (uses tsc incremental cache, for example)

## TODO

* extract testlib into external dependency. (tracked by #1233)

---

Ava results from `npm test`

```
  79 tests passed
------------------------------|---------|----------|---------|---------|--------------------------------------------------------------------------------------------------------
File                          | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
------------------------------|---------|----------|---------|---------|--------------------------------------------------------------------------------------------------------
All files                     |    48.1 |    69.24 |   36.97 |    50.3 |
 ts-node                      |       0 |      100 |     100 |       0 |
  esm.mjs                     |       0 |      100 |     100 |       0 | 3-7
 ts-node/dist-raw             |   76.87 |    59.02 |   71.43 |   77.69 |
  node-cjs-loader-utils.js    |   82.35 |    59.26 |     100 |   83.67 | 32,39,55-56,71-73,120
  node-errors.js              |   66.67 |      100 |      50 |   66.67 | 18
  node-internal-fs.js         |   66.67 |        0 |     100 |      75 | 10-11
  node-options.js             |   78.95 |     62.5 |     100 |   78.38 | 64-69,75-76,88,98
  node-package-json-reader.js |   84.21 |     62.5 |     100 |   84.21 | 19,36-37
  node-primordials.js         |      50 |      100 |   46.15 |      50 | 3-4,14-15,17,19,21
 ts-node/esm                  |       0 |      100 |     100 |       0 |
  transpile-only.mjs          |       0 |      100 |     100 |       0 | 3-7
 ts-node/register             |      25 |      100 |     100 |      25 |
  files.js                    |       0 |      100 |     100 |       0 | 1
  index.js                    |     100 |      100 |     100 |     100 |
  transpile-only.js           |       0 |      100 |     100 |       0 | 1
  type-check.js               |       0 |      100 |     100 |       0 | 1
 ts-node/src                  |   45.45 |    70.44 |   34.04 |   47.64 |
  bin-cwd.ts                  |     100 |      100 |     100 |     100 |
  bin-script-deprecated.ts    |       0 |      100 |     100 |       0 | 3-10
  bin-script.ts               |     100 |      100 |     100 |     100 |
  bin-transpile.ts            |       0 |      100 |     100 |       0 | 3-5
  bin.ts                      |   92.94 |    92.42 |    87.5 |    93.9 | 222,225,243-244,251
  esm.ts                      |   91.67 |    82.61 |     100 |   91.11 | 33,44,50,91
  index.spec.ts               |       0 |        0 |       0 |       0 | 1-879
  index.ts                    |   85.42 |    78.25 |   84.62 |   87.37 | 104,108-111,591,684-688,724-725,826,831,875-881,963,968-972,984-999,1006-1008,1014,1032,1198,1310-1324
  repl.ts                     |   82.35 |     69.7 |   90.48 |   84.96 | 86-87,93-101,205-209,240-241,262-263,279,323
  testlib.ts                  |       0 |        0 |       0 |       0 | 1-157
------------------------------|---------|----------|---------|---------|--------------------------------------------------------------------------------------------------------

/d/Personal-dev/@TypeStrong/ts-node ab/test-with-ava* ⇡ 1m 50s
```

Mocha results from `npm test`

```
  79 passing (2m)

------------------------------|---------|----------|---------|---------|--------------------------------------------------------------------------------------------------------
File                          | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
------------------------------|---------|----------|---------|---------|--------------------------------------------------------------------------------------------------------
All files                     |   50.65 |    72.09 |   39.76 |   52.98 |
 ts-node                      |       0 |      100 |     100 |       0 |
  esm.mjs                     |       0 |      100 |     100 |       0 | 3-7
 ts-node/dist-raw             |   76.87 |    59.02 |   71.43 |   77.69 |
  node-cjs-loader-utils.js    |   82.35 |    59.26 |     100 |   83.67 | 32,39,55-56,71-73,120
  node-errors.js              |   66.67 |      100 |      50 |   66.67 | 18
  node-internal-fs.js         |   66.67 |        0 |     100 |      75 | 10-11
  node-options.js             |   78.95 |     62.5 |     100 |   78.38 | 64-69,75-76,88,98
  node-package-json-reader.js |   84.21 |     62.5 |     100 |   84.21 | 19,36-37
  node-primordials.js         |      50 |      100 |   46.15 |      50 | 3-4,14-15,17,19,21
 ts-node/esm                  |       0 |      100 |     100 |       0 |
  transpile-only.mjs          |       0 |      100 |     100 |       0 | 3-7
 ts-node/register             |      25 |      100 |     100 |      25 |
  files.js                    |       0 |      100 |     100 |       0 | 1
  index.js                    |     100 |      100 |     100 |     100 |
  transpile-only.js           |       0 |      100 |     100 |       0 | 1
  type-check.js               |       0 |      100 |     100 |       0 | 1
 ts-node/src                  |   48.14 |    73.69 |   36.84 |   50.49 |
  bin-cwd.ts                  |     100 |      100 |     100 |     100 |
  bin-script-deprecated.ts    |       0 |      100 |     100 |       0 | 3-10
  bin-script.ts               |     100 |      100 |     100 |     100 |
  bin-transpile.ts            |       0 |      100 |     100 |       0 | 3-5
  bin.ts                      |   92.94 |    92.42 |    87.5 |    93.9 | 222,225,243-244,251
  esm.ts                      |   91.67 |    82.61 |     100 |   91.11 | 33,44,50,91
  index.spec.ts               |       0 |        0 |       0 |       0 | 1-877
  index.ts                    |   85.42 |    78.25 |   84.62 |   87.37 | 103,107-110,590,683-687,723-724,825,830,874-880,962,967-971,983-998,1005-1007,1013,1031,1197,1309-1323
  repl.ts                     |   82.35 |     69.7 |   90.48 |   84.96 | 86-87,93-101,205-209,240-241,262-263,279,323
------------------------------|---------|----------|---------|---------|--------------------------------------------------------------------------------------------------------

/d/Personal-dev/@TypeStrong/ts-node-master master 2m 47s
```